Add full describe call path and modify APIs
This commit does not include any reachable code paths.
We are opting to reduce number of needed API calls which we originally had before
with multiple calls to fetch provisioning artifacts and launch paths separately in order to
handle conerns over pagination. Instead we will just call describe product which returns
everything in the response body with minimal latency differences.

This also sets up adding both ability to describe a single product, as well as describing
all available products, which we have default paths for selecting the PA and launch path for.